// src/features/Scenario/types.ts
export interface BaseData {
  created_at: Timestamp;
  updated_at: Timestamp;
}
// ...
export interface Field {
  id: number;
  name: string;
  base_data: BaseData;
}

export interface Scenario {
  id: number;
  name: string;
  description: string;
  base_data: BaseData;
  rating: number;
  fields: Field[];
  total_participant: number;
  questions: Question[];
}

export interface Question {
  id: number;
  criteria: string;
  hint: string;
  content: string;
  base_data: BaseData;
}
// ...
export interface ScenarioQuestion {
  criteria: string;
  hint: string;
  content: string;
}